refactor: entity, dto cast 작업 부분 merge
