+++ src/game/entity/game.channel.entity.ts
@@ -16,13 +16,13 @@ import {
  * @description 게임에 대한 엔터티
  *
  * @param {number} id - 게임의 고유 ID
+ * @param {number} creator_id - 게임의 생성자 ID
+ * @param {string} creator_avatar - 게임의 생성자 프로필 사진 Base64 값
  * @param {string} title - 게임의 제목
  * @param {GameChannelPolicy} game_channel_policy - 게임의 공개 여부 (PUBLIC/PRIVATE)
  * @param {GameType} game_type - 게임의 타입 (SPEED/BOUNCE)
  * @param {GameMode} game_mode - 게임의 모드 (NORMAL/SPEED)
  * @param {GameStatus} game_status - 게임의 상태 (WAITING/PLAYING/ENDED)
- * @param {number} creator_id - 게임의 생성자 ID
- * @param {string} creator_avatar - 게임의 생성자 프로필 사진 Base64 값
  * @param {number} cur_user - 게임의 현재 유저 수
  * @param {number} max_user - 게임의 최대 유저 수
  * @param {Date} created_at - 게임의 생성 시간
