+++ src/game/game.gateway.ts
@@ -12,7 +12,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { Game } from "./entity/game.entity";
 import { Redis } from "ioredis";
 import { GameUserRole, GameStatus, GameChannelPolicy } from "./enum/game.enum";
-import { GamePlayer } from "./entity/game.players.entity";
+import { GamePlayer } from "./entity/game.player.entity";
 import { JWTAuthGuard } from "src/auth/jwt/jwtAuth.guard";
 import { format } from "date-fns";
 import { UserService } from "src/user/user.service";
